separate the processing steps in error cost computation
steps:
1. process error cost with example set;
2. process error cost with verifying result;
3. insert counterexamples into example set

Still NEED TO CHECK whether the counterexample has already been in the example set.
For examples, if the example set is {(10,15)}, the output of synth is 10, then (10,15)is also an counterexample, which shows that example chek cannot ensure the same counterexample generation, since the example that synth fails in the example set is a counterexample.